move getmore command test
